refactor: fix compiler warnings and improve code quality
- Remove redundant type cast in class_file_parser.rs
- Fix unused variable warning in main.rs
- Use string interpolation in java_class_generator.rs panic message

+++ src/jvm/class_file_parser.rs
@@ -137,10 +137,7 @@ impl ClassFileParser {
                 }
                 _ => {
                     // Unknown constant pool tag
-                    return Err(RuntimeError::UnknownConstantPoolTag {
-                        tag,
-                        index: i as u16,
-                    });
+                    return Err(RuntimeError::UnknownConstantPoolTag { tag, index: i });
                 }
             }
             i += 1;
